Moved more stuff over to web infra
