+++ src/HttPlaceholder.WebInfrastructure.Tests/Implementations/ClientDataResolverFacts.cs
@@ -2,7 +2,7 @@ using HttPlaceholder.Application.Configuration;
 using HttPlaceholder.Web.Shared.Infrastructure.Web;
 using Microsoft.AspNetCore.Http;
 
-namespace HttPlaceholder.Web.Shared.Tests.Infrastructure.Web;
+namespace HttPlaceholder.WebInfrastructure.Tests.Implementations;
 
 [TestClass]
 public class ClientIpResolverFacts
