+++ src/HttPlaceholder.WebInfrastructure/Implementations/ClientDataResolver.cs
