+++ src/HttPlaceholder.WebInfrastructure/Properties/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("HttPlaceholder.WebInfrastructure.Tests")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
